+++ src/candlewick/core/DepthAndShadowPass.cpp
@@ -8,33 +8,42 @@
 
 namespace candlewick {
 
-static SDL_GPUGraphicsPipeline *
+static GraphicsPipeline
 create_depth_pass_pipeline(const Device &device, const MeshLayout &layout,
                            SDL_GPUTextureFormat format,
                            const DepthPass::Config config) {
   auto vertexShader = Shader::fromMetadata(device, "ShadowCast.vert");
   auto fragmentShader = Shader::fromMetadata(device, "ShadowCast.frag");
-  SDL_GPUGraphicsPipelineCreateInfo pipeline_desc{
-      .vertex_shader = vertexShader,
-      .fragment_shader = fragmentShader,
-      .vertex_input_state = layout.toVertexInputState(),
-      .primitive_type = SDL_GPU_PRIMITIVETYPE_TRIANGLELIST,
-      .rasterizer_state{
-          .fill_mode = SDL_GPU_FILLMODE_FILL,
-          .cull_mode = config.cull_mode,
-          .depth_bias_constant_factor = config.depth_bias_constant_factor,
-          .depth_bias_slope_factor = config.depth_bias_slope_factor,
-          .enable_depth_bias = config.enable_depth_bias,
-          .enable_depth_clip = config.enable_depth_clip},
-      .depth_stencil_state{.compare_op = SDL_GPU_COMPAREOP_LESS,
-                           .enable_depth_test = true,
-                           .enable_depth_write = true},
-      .target_info{.color_target_descriptions = nullptr,
-                   .num_color_targets = 0,
-                   .depth_stencil_format = format,
-                   .has_depth_stencil_target = true},
-  };
-  return SDL_CreateGPUGraphicsPipeline(device, &pipeline_desc);
+  return GraphicsPipeline(
+      device,
+      {
+          .vertex_shader = vertexShader,
+          .fragment_shader = fragmentShader,
+          .vertex_input_state = layout.toVertexInputState(),
+          .primitive_type = SDL_GPU_PRIMITIVETYPE_TRIANGLELIST,
+          .rasterizer_state{
+              .fill_mode = SDL_GPU_FILLMODE_FILL,
+              .cull_mode = config.cull_mode,
+              .depth_bias_constant_factor = config.depth_bias_constant_factor,
+              .depth_bias_slope_factor = config.depth_bias_slope_factor,
+              .enable_depth_bias = config.enable_depth_bias,
+              .enable_depth_clip = config.enable_depth_clip,
+          },
+          .multisample_state{},
+          .depth_stencil_state{
+              .compare_op = SDL_GPU_COMPAREOP_LESS,
+              .enable_depth_test = true,
+              .enable_depth_write = true,
+          },
+          .target_info{
+              .color_target_descriptions = nullptr,
+              .num_color_targets = 0,
+              .depth_stencil_format = format,
+              .has_depth_stencil_target = true,
+          },
+          .props = 0,
+      },
+      config.pipeline_name);
 }
 
 DepthPass::DepthPass(const Device &device, const MeshLayout &layout,
@@ -42,32 +51,23 @@ DepthPass::DepthPass(const Device &device, const MeshLayout &layout,
                      const Config &config)
     : _device(device), depthTexture(depth_texture) {
   pipeline = create_depth_pass_pipeline(device, layout, format, config);
-  if (!pipeline)
-    terminate_with_message("Failed to create graphics pipeline: {:s}.",
-                           SDL_GetError());
 }
-
-DepthPass::DepthPass(const Device &device, const MeshLayout &layout,
-                     const Texture &texture, const Config &config)
-    : DepthPass(device, layout, texture, texture.format(), config) {}
 
 DepthPass::DepthPass(DepthPass &&other) noexcept
     : _device(other._device)
     , depthTexture(other.depthTexture)
-    , pipeline(other.pipeline) {
+    , pipeline(std::move(other.pipeline)) {
   other._device = nullptr;
   other.depthTexture = nullptr;
-  other.pipeline = nullptr;
 }
 
 DepthPass &DepthPass::operator=(DepthPass &&other) noexcept {
   if (this != &other) {
     _device = other._device;
     depthTexture = other.depthTexture;
-    pipeline = other.pipeline;
+    pipeline = std::move(other.pipeline);
     other._device = nullptr;
     other.depthTexture = nullptr;
-    other.pipeline = nullptr;
   }
   return *this;
 }
@@ -85,7 +85,7 @@ void DepthPass::render(CommandBuffer &command_buffer, const Mat4f &viewProj,
 
   SDL_GPURenderPass *render_pass =
       SDL_BeginGPURenderPass(command_buffer, nullptr, 0, &depth_info);
-  SDL_BindGPUGraphicsPipeline(render_pass, pipeline);
+  pipeline.bind(render_pass);
 
   for (auto &[mesh, tr] : castables) {
     assert(validateMesh(mesh));
@@ -105,15 +105,10 @@ gpuViewportFromAtlasRegion(const ShadowMapPass::AtlasRegion &reg) {
 };
 
 void DepthPass::release() noexcept {
-  if (_device) {
-    if (pipeline) {
-      SDL_ReleaseGPUGraphicsPipeline(_device, pipeline);
-      pipeline = nullptr;
-    }
-    if (depthTexture)
-      depthTexture = nullptr;
+  if (_device && depthTexture) {
+    depthTexture = nullptr;
   }
-  _device = nullptr;
+  pipeline.release();
 }
 
 void ShadowMapPass::configureAtlasRegions(const Config &config) {
@@ -134,7 +129,7 @@ void ShadowMapPass::configureAtlasRegions(const Config &config) {
 
 ShadowMapPass::ShadowMapPass(const Device &device, const MeshLayout &layout,
                              SDL_GPUTextureFormat format, const Config &config)
-    : _device(device), _numLights(config.numLights) {
+    : m_device(device), m_numLights(config.numLights) {
 
   const Uint32 atlasWidth = config.numLights * config.width;
   const Uint32 atlasHeight = config.height;
@@ -163,10 +158,8 @@ ShadowMapPass::ShadowMapPass(const Device &device, const MeshLayout &layout,
                                             config.depth_bias_slope_factor,
                                             config.enable_depth_bias,
                                             config.enable_depth_clip,
+                                            nullptr,
                                         });
-  if (!pipeline)
-    terminate_with_message("Failed to create shadow cast pipeline: {:s}.",
-                           SDL_GetError());
 
   SDL_GPUSamplerCreateInfo sample_desc{
       .min_filter = SDL_GPU_FILTER_LINEAR,
@@ -182,46 +175,41 @@ ShadowMapPass::ShadowMapPass(const Device &device, const MeshLayout &layout,
 }
 
 ShadowMapPass::ShadowMapPass(ShadowMapPass &&other) noexcept
-    : _device(other._device)
-    , _numLights(other._numLights)
+    : m_device(other.m_device)
+    , m_numLights(other.m_numLights)
     , shadowMap(std::move(other.shadowMap))
-    , pipeline(other.pipeline)
+    , pipeline(std::move(other.pipeline))
     , sampler(other.sampler)
     , cam(std::move(other.cam))
     , regions(std::move(other.regions)) {
-  other._device = nullptr;
-  other.pipeline = nullptr;
+  other.m_device = nullptr;
   other.sampler = nullptr;
 }
 
 ShadowMapPass &ShadowMapPass::operator=(ShadowMapPass &&other) noexcept {
-  _device = other._device;
-  _numLights = other._numLights;
+  m_device = other.m_device;
+  m_numLights = other.m_numLights;
   shadowMap = std::move(other.shadowMap);
   sampler = other.sampler;
-  pipeline = other.pipeline;
+  pipeline = std::move(other.pipeline);
   cam = std::move(other.cam);
   regions = std::move(other.regions);
 
-  other._device = nullptr;
-  other.pipeline = nullptr;
+  other.m_device = nullptr;
   other.sampler = nullptr;
   return *this;
 }
 
 void ShadowMapPass::release() noexcept {
-  if (_device) {
+  if (m_device) {
     if (sampler) {
-      SDL_ReleaseGPUSampler(_device, sampler);
+      SDL_ReleaseGPUSampler(m_device, sampler);
     }
     sampler = nullptr;
-    if (pipeline) {
-      SDL_ReleaseGPUGraphicsPipeline(_device, pipeline);
-    }
-    pipeline = nullptr;
   }
+  pipeline.release();
   shadowMap.destroy();
-  _device = nullptr;
+  m_device = nullptr;
 }
 
 void ShadowMapPass::render(CommandBuffer &command_buffer,
@@ -237,7 +225,7 @@ void ShadowMapPass::render(CommandBuffer &command_buffer,
 
   SDL_GPURenderPass *render_pass =
       SDL_BeginGPURenderPass(command_buffer, nullptr, 0, &depth_info);
-  SDL_BindGPUGraphicsPipeline(render_pass, pipeline);
+  pipeline.bind(render_pass);
 
   for (size_t i = 0; i < numLights(); i++) {
     SDL_GPUViewport vp = gpuViewportFromAtlasRegion(regions[i]);